+++ packages/render-composer/src/effects/SMAAEffect.tsx
@@ -1,5 +1,5 @@
 import * as PP from "postprocessing"
-import { usePostProcessingEffect } from "./lib/usePostProcessingEffect"
+import { usePostProcessingEffect } from "../lib/usePostProcessingEffect"
 
 export const SMAAEffect = (
   props: ConstructorParameters<typeof PP.SMAAEffect>[0]
