+++ packages/render-composer/src/effects/VignetteEffect.tsx
@@ -1,5 +1,5 @@
 import * as PP from "postprocessing"
-import { usePostProcessingEffect } from "./lib/usePostProcessingEffect"
+import { usePostProcessingEffect } from "../lib/usePostProcessingEffect"
 
 export const VignetteEffect = (
   props: ConstructorParameters<typeof PP.VignetteEffect>[0]
