+++ packages/render-composer/src/index.ts
@@ -1,9 +1,6 @@
-export * from "./passes"
 export * from "./Canvas"
 export * from "./EffectComposer"
-export * from "./GodRaysEffect"
+export * from "./effects"
 export * from "./Layers"
+export * from "./passes"
 export * from "./RenderPipeline"
-export * from "./SelectiveBloomEffect"
-export * from "./SMAAEffect"
-export * from "./VignetteEffect"
